deletion confirmation dialog functions

+++ pages/workouts/[id].jsx
@@ -3,7 +3,9 @@ import { useRouter } from "next/router";
 import styled from "styled-components";
 import useSWR from "swr";
 import { collectMuscleGroups } from "@/lib/utils";
+import { useState } from "react";
 export default function WorkoutDetailsPage() {
+  const [isDeleting, setIsDeleting] = useState(false);
   const router = useRouter();
   const id = router.query.id;
   const apiUrl = `/api/workouts/${id}`;
@@ -22,6 +24,9 @@ export default function WorkoutDetailsPage() {
   const muscleGroupsInWorkout = collectMuscleGroups(exercises);
 
   const handleDelete = async () => {
+    if (!isDeleting) {
+      return setIsDeleting(true);
+    }
     const response = await fetch(apiUrl, {
       method: "DELETE",
     });
@@ -59,9 +64,20 @@ export default function WorkoutDetailsPage() {
       >
         update
       </button>
-      <button type="button" onClick={handleDelete}>
-        Delete
-      </button>
+      {isDeleting ? (
+        <>
+          <button type="button" onClick={() => setIsDeleting(false)}>
+            Cancel Delete
+          </button>
+          <button type="button" onClick={handleDelete}>
+            Confirm Deletion
+          </button>
+        </>
+      ) : (
+        <button type="button" onClick={handleDelete}>
+          Delete
+        </button>
+      )}
       <div>
         <Link href={`/`}>home</Link>
         <Link href={`/workouts`}>Workouts</Link>
